Organize modules under app directory
