+++ app/vista_centro_accion.py
